+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - Update Font Awesome icons
+- Add a close button to modal
 
 ## [0.1.6] - 2019-02-22
 ### Added